requirements.txt
requests==2.25.1
mysql-connector-python==8.0.25
